requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.3.1
pandas>=2.0.0
matplotlib>=3.7.1
wordcloud>=1.9.2
textblob>=0.17.1